Move to using override mechanism to hook the common globals and require behavior. Must define console overrides in Ti.API overrides as well or we run into a dependency issue accessing the console object before it's overriden.
