+++ CHANGELOG.md
@@ -1,12 +1,10 @@
-## 1.1.1
-
-### Bug Fixes
-* Combined the require-provider and common-globals plugins with the ti-api-provider
-
 ## 1.1.0
 
 ## New Features
 * Implemented a proper CLI hook
+
+### Bug Fixes
+* Combined the require-provider and common-globals plugins with the ti-api-provider
 
 ## 1.0.1
